requirements.txt
pytest==3.8.2
pytest-allure-adaptor==1.7.10
pytest-chinese-nodeid==1.2
pytest-html==1.19.0
PyYAML==3.12
requests==2.18.4
selenium==3.12.0
pocoui==1.0.70
airtest==1.0.24
